Jquery is a special case (See note below)
Notes from @eomuraliev

Incuding jQuery is a special case. If you read the [jQuery section in the  RequireJS docs](http://requirejs.org/docs/jquery.html) you will see a code example:
```
// THIS WILL NOT WORK
    define(['lib/jquery'], function ($) {...});
```

To be honest, reading over the documentation, the phrasing does not make it very clear, and examples kind of jump around.

Basically, in your case, to make it work, you gotta put the following at the top of your `app.js` file:
```
requirejs.config({
    "paths": {
      "jquery": "lib/jquery"
    }
});
```

and then when you require jquery in your other scripts, just require it simply using `jquery`, and it will load fine. 

The reason for using just `jquery` and not `lib/jquery` is stated in the documentation: 

> So to reiterate, you will likely get an error if you refer to jQuery with another module name, like 'lib/jquery' ... since jQuery registers itself with the name of 'jquery' and not 'lib/jquery'. In general, explicitly naming modules in the define() call are discouraged, but jQuery has some special constraints.

Let me know if it works for you.